CLN: fix typos. Add new callbacks module
